Add initial Jinja templates
Added Jinja templates required to generate first draft of a text chapter page, including a general template that should be used as the base for all pages, and an extending template specific to text chapter pages. Also added templates for the top bar and sidebar to be used accross most/all site pages.