Implement hotfix to get current module + section
Previous code simply lumped the currrent module or section (whichever
one did not have a surrounding anchor tag) with the others during
extraction, meaning that information was lost when trying to determine
what section or module a page belonged to. This hotfix changes the
sidebar and topbar extraction code so that the module and section that a
page belongs to are returned as part of the extraction results.