Add basic text page generation functionality
Adds the barebones of generation functionality for EOT text chapter 
pages. Right now the files lack some of the planned features including 
(but not limited to): pagination, popovers for figures/images and 
references. Also presently does not include any stylesheet resources. 
Code is unlikely to run on Windows at the moment, this will be addressed 
in the next commits.

Jinja templates needed to be reworked to be usable; some syntax was 
incorrect, and the variables expected by the templates needed to be 
tweaked.