Fix bugs preventing extraction of non-part2 parts
Earlier code would detect some sidebars with only a module title and no
author or section links as having two <p> tags, and thus fail. The fix
removes any empty <p> tags so the code properly detects what type of
sidebar is being extracted from.

The code also adds a section to any modules for which the sidebar has no
links, i.e. for modules without any sections except the default. This
allows the page content, number, etc. to be properly stored, whereas
before page info was not actually extracted for a module with only one
section.

Lastly a bug fix is added for assigning page numbers to sections in
extract_full_module, as the previous code failed in the "Food Remains"
chapter on the "Plants 1983-1984" module due to a section titled
"Summary" appearing twice in the sidebar.