Fix bugs and resolve paths in extraction code
Fix bug in extracting references where extraction would fail on an
author without a corresponding ref_xx.html page for the author name, but
with a corresponding page for one of the references. Also fix a specific
case where the original references had inconsistent HTML.

Additionally, change the code for extracting text so that relative paths
in <a> tags become full paths in the extracted data, to eliminate any
inconsistencies when resolving old paths to new ones during generation.