Update code and tests to support digpro
Code previously would work with paths while including "/dig" at the
start; however, we can instead work with paths starting from "/dig/html"
or "/digpro/html", while not including the "/dig" or "/digpro". The only
material that can't be extracted then would be the old electronic dig
and the old index page, which we do not need to extract.

The new generation method assumes that generate.py will be run twice,
once for /dig and once for /digpro. That is, both versions of the site
will not be generated using the same "index" object.