Fix bugs in references, excavations, and home page
This commit spreads the "A-G", "H-R", and "S-Z" pages in References
across three modules, whereas before they were three sections in a
single module, in order to correct the issue of "References" appearing
twice in the sidebar. A CSS style is also added to fix indentation for
each reference.

This commit also fixes sorting of contexts in the Excavations chapter
sidebar; contexts are now sorted by number, e.g. 1, 2, 3, 4, 5, ...,
whereas before they were either random on MacOS due to how the data
extraction was implemented, or ordered 1, 10, 11, ..., 19, 2, 20, etc.

The spelling of "Archaeology" in the home page is fixed, and initials
are added to some of the authors' names.

Finally, remove exc_elem_artifacts.html.jinja as it is no longer used.