Fix more styling issues and other bugs
- Fix extraction of author initials at the end of both Prefaces in the
Introduction chapter.
- Fix the updating of image links for some Archaeology Primer pages.
- Make max-width for auto-collapsing the sidebar 767px, not 768px.
- Add Bootstrap styling to tables in the Data Downloads chapter.
- Make more sidebar fixes through removing old CSS rules. max-width on
Bootstrap cols are removed so cols can expand according to default
settings, and white-space: nowrap is removed so sidebar elements will
not overflow improperly on small screens.
- Improve styling of the interactive site grid (page 5) and changeable
images (pages 13, 14, and 23) in the Archaeology Primer.
- Improve look on mobile by restoring padding-right on .container-fluid;
removing "overflow-y: scroll;" from the main content col to get rid of
the extra scrollbar; setting size limits on images in the archaeology
primer and allowing figcaptions to wrap in order to ensure the sidebar
and main content cols can re-size properly; and adjusting font-size on
tables in Appendix A and making tables scroll horizontally in both
Appendix A and B.