Sidebar changes and other bug fixes
- Changed sidebar generation to use macros. No more maintaining two 
separate files for the sidebar.

- Sidebar now uses col-lg-3 and col-md-4 instead of col-md-3. This 
solves vertical overflow issues that were popping up in some browsers 
slightly wider than the Bootstrap md breakpoint. It also just generally 
looks better for the sidebar to fill 33% of the screen at that width. 
Changed script for Appendix B to reflect different sidebar sizes for md 
and lg.

- Styles: #toc .nav no longer has 'display: block', sidebar-module, 
-section, and -subsection all have full width. Also fixed a bottom 
margin appearing beneath the author <p> in the sidebar.

- "Related Elements" -> "Related Contexts" on the excavation pages.

- Removed redundant stylesheet listings.