Add favicons and Excavations map tooltip
- Add favicons generated from realfavicongenerator.net. Note that the
browserconfig.xml and site.webmanifest files have had their contents
changed to use a relative link rather than an absolute path.
- Add functionality for displaying a tooltip with the name of whichever
context the user's mouse is hovering over when visiting the map in
the "Excavations" chapter on desktop. Whether the map will disable the
tooltip on mobile, as it is intended to, still needs to be tested.
- Corrected spelling of "Occaneechi" in index.html.jinja.