Modify pagination display and Appendix A sidebar
- Sort sidebar of Appendix A into Features, Squares, and Structures, as
well as numerically sort the contexts in each sidebar category. That is,
the sidebar now displays Feature 1, 2, 3, ... instead of Feature 1, 10,
11, ... However, the "Previous" and "Next" page buttons still use the
page numbers from the old site, so that clicking the "Next" button on
Feature 1 still brings the user to Feature 10. This is an intentional
compromise.
- Make the first and last pages of distinct chapters from a pagination
standpoint (i.e. Getting Started, Archaeology Primer,
Introduction+Contents, etc.) link between one another when using the
"Previous" and "Next" buttons.
- Change "Previous" and "Next" pagination buttons to use Bootstrap arrow
icons. The svg tags are taken directly from the Bootstrap website, with
the exception that a "vertical-align:bottom;" rule was added to avoid
issues with the arrow buttons sometimes being taller than the page
number tile.
- Make clear that the page number is clickable when hovering over it.