Fix index.html button sizing
The buttons on the index.html page used to have clickable areas that
were longer than the actual, displayed button. This fix makes it so the
clickable area of the button and the displayed part are the same.
However, as Internet Explorer does not support "width: max-content;",
the buttons will be of different width on that browser vs Chrome, etc.